refactor: Github OAuth Exchange 로직 추상화
