+++ backend/src/main/java/com/rssmanager/auth/service/OAuthExchangeService.java
@@ -1,0 +1,9 @@
+package com.rssmanager.auth.service;
+
+import com.rssmanager.auth.controller.dto.LoginRequest;
+import com.rssmanager.member.domain.Member;
+
+@FunctionalInterface
+public interface OAuthExchangeService {
+    Member fetchMember(LoginRequest loginRequest);
+}
