+++ backend/src/main/java/com/rssmanager/util/SessionManager.java
@@ -4,11 +4,11 @@ import com.rssmanager.member.domain.Member;
 import javax.servlet.http.HttpSession;
 
 public interface SessionManager {
-    boolean isLoggedIn();
-
     HttpSession login(Member member);
 
     <T> T getAttribute(String key);
 
     void invalidate();
+
+    Member getLoginMember();
 }
